use CatalogTableToolbar in all Catalogs, move title
