+++ plugins/catalog/src/components/CatalogTable/CatalogTable.tsx
@@ -38,8 +38,6 @@ import Typography from '@material-ui/core/Typography';
 import { visuallyHidden } from '@mui/utils';
 import Edit from '@material-ui/icons/Edit';
 import OpenInNew from '@material-ui/icons/OpenInNew';
-import { capitalize } from 'lodash';
-import pluralize from 'pluralize';
 import React, { ReactNode, useMemo } from 'react';
 import { columnFactories } from './columns';
 import { CatalogTableColumnsFunc, CatalogTableRow } from './types';
@@ -49,6 +47,7 @@ import { defaultCatalogTableColumnsFunc } from './defaultCatalogTableColumnsFunc
 import { useTranslationRef } from '@backstage/core-plugin-api/alpha';
 import { catalogTranslationRef } from '../../alpha/translation';
 import { FavoriteToggleIcon } from '@backstage/core-components';
+import { CatalogTableToolbar } from './CatalogTableToolbar';
 
 /**
  * Props for {@link CatalogTable}.
@@ -84,15 +83,8 @@ export const CatalogTable = (props: CatalogTableProps) => {
   const { isStarredEntity, toggleStarredEntity } = useStarredEntities();
   const entityListContext = useEntityList();
 
-  const {
-    loading,
-    error,
-    entities,
-    filters,
-    pageInfo,
-    totalItems,
-    paginationMode,
-  } = entityListContext;
+  const { loading, error, entities, pageInfo, paginationMode } =
+    entityListContext;
 
   const tableColumns = useMemo(
     () =>
@@ -168,20 +160,6 @@ export const CatalogTable = (props: CatalogTableProps) => {
     },
   ];
 
-  const currentKind = filters.kind?.value || '';
-  const currentType = filters.type?.value || '';
-  const currentCount = typeof totalItems === 'number' ? `(${totalItems})` : '';
-  // TODO(timbonicus): remove the title from the CatalogTable once using EntitySearchBar
-  const titlePreamble = capitalize(filters.user?.value ?? 'all');
-  const title = [
-    titlePreamble,
-    currentType,
-    pluralize(currentKind),
-    currentCount,
-  ]
-    .filter(s => s)
-    .join(' ');
-
   const actions = props.actions || defaultActions;
   const options = {
     actionsColumnIndex: -1,
@@ -197,7 +175,6 @@ export const CatalogTable = (props: CatalogTableProps) => {
         columns={tableColumns}
         emptyContent={emptyContent}
         isLoading={loading}
-        title={title}
         actions={actions}
         subtitle={subtitle}
         options={options}
@@ -212,7 +189,6 @@ export const CatalogTable = (props: CatalogTableProps) => {
         columns={tableColumns}
         emptyContent={emptyContent}
         isLoading={loading}
-        title={title}
         actions={actions}
         subtitle={subtitle}
         options={options}
@@ -235,11 +211,13 @@ export const CatalogTable = (props: CatalogTableProps) => {
         pageSizeOptions: [20, 50, 100],
         ...options,
       }}
-      title={title}
       data={rows}
       actions={actions}
       subtitle={subtitle}
       emptyContent={emptyContent}
+      components={{
+        Toolbar: CatalogTableToolbar,
+      }}
     />
   );
 };
